requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
oci==2.163.1
Jinja2==3.1.4
